Prepare refactoring to externalize reusable stuff across modules.
