+++ modules/ActivationModule/nbproject/configurations.xml
@@ -12,6 +12,7 @@
       <itemPath>LockNotificationTask.hh</itemPath>
       <itemPath>LowPowerLed.hh</itemPath>
       <itemPath>MatrixKeypad.hh</itemPath>
+      <itemPath>NetworkUtils.hh</itemPath>
       <itemPath>PingTask.hh</itemPath>
       <itemPath>Pins.hh</itemPath>
       <itemPath>RTTUtils.hh</itemPath>
@@ -26,6 +27,9 @@
                    displayName="Source Files"
                    projectFiles="true">
       <itemPath>ActivationModule.cpp</itemPath>
+      <itemPath>ActivationNetwork.cpp</itemPath>
+      <itemPath>Cipher.cpp</itemPath>
+      <itemPath>NetworkUtils.cpp</itemPath>
     </logicalFolder>
     <logicalFolder name="TestFiles"
                    displayName="Test Files"
@@ -92,9 +96,13 @@
       </item>
       <item path="ActivationModule.cpp" ex="false" tool="1" flavor2="0">
       </item>
+      <item path="ActivationNetwork.cpp" ex="false" tool="1" flavor2="0">
+      </item>
       <item path="ActivationNetwork.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="ArrayAlloc.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="Cipher.cpp" ex="false" tool="1" flavor2="0">
       </item>
       <item path="Cipher.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -105,6 +113,10 @@
       <item path="LowPowerLed.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="MatrixKeypad.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="NetworkUtils.cpp" ex="false" tool="1" flavor2="0">
+      </item>
+      <item path="NetworkUtils.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="PingTask.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -173,9 +185,13 @@
       </item>
       <item path="ActivationModule.cpp" ex="false" tool="1" flavor2="0">
       </item>
+      <item path="ActivationNetwork.cpp" ex="false" tool="1" flavor2="0">
+      </item>
       <item path="ActivationNetwork.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="ArrayAlloc.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="Cipher.cpp" ex="false" tool="1" flavor2="0">
       </item>
       <item path="Cipher.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -186,6 +202,10 @@
       <item path="LowPowerLed.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="MatrixKeypad.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="NetworkUtils.cpp" ex="false" tool="1" flavor2="0">
+      </item>
+      <item path="NetworkUtils.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="PingTask.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -246,9 +266,13 @@
       </item>
       <item path="ActivationModule.cpp" ex="false" tool="1" flavor2="0">
       </item>
+      <item path="ActivationNetwork.cpp" ex="false" tool="1" flavor2="0">
+      </item>
       <item path="ActivationNetwork.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="ArrayAlloc.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="Cipher.cpp" ex="false" tool="1" flavor2="0">
       </item>
       <item path="Cipher.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -259,6 +283,10 @@
       <item path="LowPowerLed.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="MatrixKeypad.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="NetworkUtils.cpp" ex="false" tool="1" flavor2="0">
+      </item>
+      <item path="NetworkUtils.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="PingTask.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -325,9 +353,13 @@
       </item>
       <item path="ActivationModule.cpp" ex="false" tool="1" flavor2="0">
       </item>
+      <item path="ActivationNetwork.cpp" ex="false" tool="1" flavor2="0">
+      </item>
       <item path="ActivationNetwork.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="ArrayAlloc.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="Cipher.cpp" ex="false" tool="1" flavor2="0">
       </item>
       <item path="Cipher.hh" ex="false" tool="3" flavor2="0">
       </item>
@@ -338,6 +370,10 @@
       <item path="LowPowerLed.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="MatrixKeypad.hh" ex="false" tool="3" flavor2="0">
+      </item>
+      <item path="NetworkUtils.cpp" ex="false" tool="1" flavor2="0">
+      </item>
+      <item path="NetworkUtils.hh" ex="false" tool="3" flavor2="0">
       </item>
       <item path="PingTask.hh" ex="false" tool="3" flavor2="0">
       </item>
